pyproject.toml
[project]
name = "django-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "django-rest-framework>=0.1.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.6",
    "python-dotenv>=1.1.0",
]
